Suppress the denials related to persist.vendor.service.bdroid.bdaddr
Currently redbull reads the Bluetooth Device Address from "ro.vendor.bt.bdaddr_path",
which is a CDT kernel node.
Only when redbull cannot get the BDA from CDT, it generates a random address and set
it to "persist.vendor.service.bdroid.bdaddr" as a fallback.

This denial is not what we expect to happen. It is only shown on the abnormal device.
Now suppress this denial since we do not need it.

https://android.googlesource.com/device/google/redbull-sepolicy/+/6aeab393fe29dcf972e4b61162b071622af9fbd5